Bug 11626 - Cannot receive serial issues after subscription expiry date
Some of our serials subscriptions were not renewed after 31/12/2013 but I
cannot receive any of the outstanding 2013 issues unless I renew the
subscription record (a couple of publishers are about 3 months late with their
printing/postage schedules compared to the subscription/cover dates).

We should still be able to receive late issues after the subscription expiry
date (the date published for these issues would in most cases be on or before
the expiry date).

THIS PATCH:

Changes expiration date checking for each serial to be calculated from the
individual Serial's serial.publicationdate instead of the Serial's
Subscription's expirationdate.
Thus you can still receive late serials if their publicationdate is within
the Subscription period.

Replicating this issue in finicky, just as the Serials module is, but this
patch fixes this design flaw in the Serials-module.

REPLICATE THE ISSUE:
0. Have a Subscription with an end date.
1. Receive all the Subscription's Serials.
2. You cannot receive the last Serial of the subscription period.

TEST PLAN:
0. Have a Subscription with an end date.
1. Receive all the Subscription's Serials.
2. Receive the last Subscription of the subscription period.