Generate a new barcode with Ajax
This commit is used to generate a new barcode for Add item -form.It
will request a new barcode from the file that generates barcodes
and prints it out in JSON. It also includes some changes to the
ValueBuilder.pm to make it ask the JSON barcode from barcode_ajax.pl.
NOTE: This commit only applies to barcodes generated in annual-format.