KD#1202, Update record sublocation and notforloan status
This commit will automatically set notforloan status to 0 for records
that are checked in and have their notforloan status set as < 0.
It will also change records sub_location to null if it is set as
TIGHT or EXHIBIT. For this update to work you must change notforloan
values 2 and 3 to negative (-2 and -3). You should also update the
items that currently have their notforloan status set either as
2 or 3.