KD#377 - Add database table locks to payments to prevent data inconsistency
This patch adds locks to the payments_transactions table during CompletePayment()
subroutine. We want to prevent a case where the subroutine is called twice at the
exact same time and for those two instances to incorrectly assume that the payment
is still pending (even though it is actually either "processing" or "paid"/"cancelled").
This would lead to double-Pay events in Koha.

Also fix issue with error reporting caused by "use bignum;".

Fixes "Reference" parameter from "integer" to "string". This was a problem with CPU
simulator. The simulator sent "Reference" as "integer", and the live system sends
the parameter as "string". We want to use "string".