KD-207-1 - OKM Annual statistics. Refactored to a configurable model. Improved error logging.
Introduces a new SysPref OKM which controls how the statistics are created.

Introduces a okm_statistics_logs-table which stores log entries (errors)
from the server-side statistics generation. Thus it is much easier to react
to errors. Also included are error descriptions and some help on how to deal
with various errors.

Removed updatedatabase.pl -entry and replaced it with Atomicupdater

Added a check to see if feature "KD-257 - Precalculated Biblio data elements -table"
is deployed correctly.