KD-207-1 - OKM Annual statistics. Fix missing end single-quote character
There were issues with translations in tools.pref. One of the lines was missing
a single quotation character. This patch adds the missing character.

This patch also replaces the new line \n characters with <pre></pre> tags,
because \n was causing "Strange line at ..." error messages into the translation
update/install script. The reason was that \n was also interpreted as a new line
in the translation files, and this was something that the script could not handle.