Bug 13799 - REST API - Swagger-UI + installation
Since the Swagger-UI is under constant development and one needs to deploy
custom code to configure it to our needs, we deploy it as a separate
project to '/home/koha/swagger-ui' and link to it from
'/home/koha/kohaclone/api/v1/doc'.
This way it is accessible from the web server url '/api/v1/doc'.

INSTALLATION:

** Get the code with our modifications **
1. cd /home/koha/
2. git clone git@github.com:KohaSuomi/swagger-ui.git

** Link the swagger-ui for the web-server **
** (there actually should be a working link already, so dont overwrite it) **
3. cd /home/koha/kohaclone/api/v1/
4. ln ../../../../swagger-ui/dist doc

** Prepare the swagger-spec for swagger-ui to display. Swagger-ui is not very good
** at resolving JSON schema references so it needs a bit of help. **
5. cd /home/koha/kohaclone/api/v1/swagger
6. perl minifySwagger.pl

7. Reap what you sow.