KD#473 - Expand material type definition in OPAC search results for some types
When searching for DVDs or other visual material, Koha shows the material type
as "Visual material". This is not very descriptive for a Borrower, and he has
to go to detailed view in order to find out the exact item type.

Expand the material type definition to show in results page the exact item type,
such as "Visual material (Blu-ray)".