KD-671 - TäTi - Centralized test cluster build tools.
To test TäTi we need to create a remote representation of TäTi for testing
purposes. This is basically a external Koha-installation which exposes a public
Z39.50 interface and the REST API and the OAI-PMH service so we can run
automated test from the TäTi-consumers.

This commit tracks the build scripts to convert a Koha-intallation to a external
test system.