KD#377 - Fix a bug with CPU error reporting
SendPayment() had a bug that caused payments with error to stay in "pending"
mode until they would time out. This patch fixes the bug with error reporting
and adds a feature to store the CPU-given error code into our database. It
wasn't stored yet because of misunderstanding - when CPU gives an error for
duplicate id or some other CPU server error, they don't send the report, but
instead only answer to the long-polling call.