Bug 15996 - Bibliographic records diffing tool
Currently there is no tool to diff differences or changes between biblios or
their different versions. This is most useful for logging who changed and what.

This feature introduces a biblio diffing module, which takes an arbitrary amount
of MARC::Records and produces a complex data structure about all the identified
changes.