Bug#1093, Generating only one message type
Adds to overdue_notices.pl a preference that you can use to generate only one type of overdue message.
There is a preferemail option and it generates email, sms or print message. Before in overdue actions if
you selected email, sms and print it created all to database. This prefers email and sms over print.
If all actions are selected the preference generates firstly email if patron has email address
then sms if patron has smsnumber and print if email or sms are not generated.