KD#377, KD#636, KD#1076 - Fix payment status on reverse status notification error
Due to payment server error, we may receive two status reports, where the first
one has successful status (paid), and then another report comes with cancelled
status (server error or something else). This makes Koha revert the payment.
This reverting feature exists, because it gives librarians the ability to revert
the transaction via POS integration.

Luckily this is a rare problem, but it happened now due to some payment server
error in online payments that we can't do anything about in Koha.

There was a small bug with setting the status of payment after reverting, and
this patch fixes it and adds a description to the transaction.