KD#1221 - Messaging preferences checked checkboxes temporarily not saved while modifying contact information
While modifying a Patron (sms or email), checked checkboxes are disabled and
unchecked if the contact information is invalid. The checks should be stored
temporarily, as a Librarian usually wants to replace old contact information
with new. However, although this feature is implemented, it was not working.

This patch fixes the issue.

Also fixes a problem in the PageObject test for messaging preferences. Syspref
EnhancedMessagingPreferences needs to be activated for the tests to succeed.