KD#1476 - Mikroväylä self-service library ban
Cronscript to collect patrons to be blocked from using self
service library, and create an xml-blocklist for Mikroväylä door
(un)locking system.

The list can be published on Koha https-server (with http-auth
configured) so that Mikroväylä can fetch the list to their
door machines and make the machines utilize it.