KD#1488 - Manage account logins on expiration/renewal
Prevent patrons from logging in when their accounts
are expired by blocking the passwords. Upon refreshing
the account, the passwords will also be re-enabled.

Enter the categorycode of patrons to manage as a
parameter (for example STAFF).

Put something like this in your crontab to have the
password management happen automatically during library
opening hours:

*/10 06-20 * * * $KOHA_PATH/misc/cronjobs/manageExpiredAccounts.pl STAFF