KD#1521 - Fix for SIP-server screen messages on checkout
When trying to checkout an item that is reserved to someone else, the
SIP-server babbles out information about the next patron in the holds
queue for the biblio in screen messages displayed on the checkout
machine.

This bug actually leaks information about the borrowers, so it
definitely needed to be fixed.