Kyyti#913 - A small fix + comment
Fix missing semicolon at the end of line 213 of ILS.pm + add comment to
MsgType.pm where the SIP alerts are disabled.

The change in MsgType.pm was never meant to be a permanent solution to
the problem it causes (sorting doesn't work with Mikroväylä check-in
machines).

The Mikroväylä check-in machines need to be fixed so that they deal
correctly with SIP alerts. In this respect the SIP-server actually worked
exactly as it should, but unfortunately we have had to stab the wrong
victim to make the sorting work.