KD#1540 - Manage zebra-indexes on tmpfs
Move zebra-index to tmpfs (ramdisk) for better peformance, especially on
a laggy IO, and keep the index on tmpfs and disk in sync. This script has
two functions:

1) It functions as an init-script that sets up tmpfs and syncs zebra-index
between the tmpfs and disk

2) It takes care of updating/re-creating the index and keeping it in sync
between the tmpfs and disk (for cron)

Update your index from crontab with something like this:
*/5 06-20 * * * $KOHA_PATH/misc/cronjobs/zebratmpfs reindex >> /home/koha/koha-dev/var/log/zebra-tmpfs 2>&1

For complete re-indexing every su-mo night, this will also sync the new index to disk:
10 00 * * 7 $KOHA_PATH/misc/cronjobs/zebratmpfs newindex >> /home/koha/koha-dev/var/log/zebra-tmpfs 2>&1

It's probably a good idea to sync the index to disk nightly:
10 00 * * 1-6 $KOHA_PATH/misc/cronjobs/zebratmpfs synctodisk >> /home/koha/koha-dev/var/log/zebra-tmpfs 2>&1

Start on Koha-server boot and sync to disk on shutdown, do (as root):
ln -s $KOHA_PATH/misc/cronjobs/zebratmpfs /etc/init.d
update-rc.d zebratmpfs defaults

Check the maximum size for tmpfs in the script (default 80G)!