KD#1576 - Print transfer slips on branch transfers
By default Koha doesn't print transfer slips when confirming branch
transfers. Presumably we will want the transfer slips in Koha-Suomi
most of the time, so this is changed in returns.tt to print the slip by
default.

Since this is just a very minute template change, maybe we can
do without adding a new syspref for this.