KD#1530 - Fix for missing $yearofpublication in EDItX notices
In cases where EDItX shipment notices are lacking publication year
information on one or more bibliographic records, the notices get
rejected. This is fixed by changing the handling of publication year
as follows:

First check MARC field 260c, use that if the information is there.
If not, check YearOfPublication tag in shipment notice and use that.

In addition publication year is no longer mandatory in EDItX notices,
so shipment notices lacking both of the above will still get
processed.